+++ src/color.test.ts
@@ -4,6 +4,7 @@ import {
 	hex2rgb,
 	hsv2Color,
 	isDark,
+	isLight,
 	randomHexColor,
 	rgb2Color,
 	isHexColorValid
@@ -132,6 +133,20 @@ describe('color', () => {
 		expect(isDark('#0000FF')).toBeTruthy();
 		expect(isDark('#FF00FF')).toBeFalsy();
 		
+		expect(isDark('#FF0000', 0.9)).toBeFalsy();
+		expect(isDark('#FFFF00', 0.9)).toBeFalsy();
+		expect(isDark('#00FF00', 0.9)).toBeFalsy();
+		expect(isDark('#00FFFF', 0.9)).toBeFalsy();
+		expect(isDark('#0000FF', 0.9)).toBeFalsy();
+		expect(isDark('#FF00FF', 0.9)).toBeFalsy();
+		
+		expect(isDark('#FF0000', 0.1)).toBeTruthy();
+		expect(isDark('#FFFF00', 0.1)).toBeFalsy(); // yellow is still one of the 10% lightest color
+		expect(isDark('#00FF00', 0.1)).toBeTruthy();
+		expect(isDark('#00FFFF', 0.1)).toBeTruthy();
+		expect(isDark('#0000FF', 0.1)).toBeTruthy();
+		expect(isDark('#FF00FF', 0.1)).toBeTruthy();
+		
 		expect(isDark(hex2rgb({ hex: '#FF0000' }))).toBeFalsy();
 		expect(isDark(hex2rgb({ hex: '#FFFF00' }))).toBeFalsy();
 		expect(isDark(hex2rgb({ hex: '#00FF00' }))).toBeFalsy();
@@ -139,6 +154,29 @@ describe('color', () => {
 		expect(isDark(hex2rgb({ hex: '#0000FF' }))).toBeTruthy();
 		expect(isDark(hex2rgb({ hex: '#FF00FF' }))).toBeFalsy();
 	});
+	
+	it('isLight', () => {
+		expect(isLight('#FF0000')).toBeTruthy();
+		expect(isLight('#FFFF00')).toBeTruthy();
+		expect(isLight('#00FF00')).toBeTruthy();
+		expect(isLight('#00FFFF')).toBeTruthy();
+		expect(isLight('#0000FF')).toBeFalsy();
+		expect(isLight('#FF00FF')).toBeTruthy();
+		
+		expect(isLight('#FF0000', 0.9)).toBeFalsy();
+		expect(isLight('#FFFF00', 0.9)).toBeTruthy(); // yellow is still one of the 10% lightest color
+		expect(isLight('#00FF00', 0.9)).toBeFalsy();
+		expect(isLight('#00FFFF', 0.9)).toBeFalsy();
+		expect(isLight('#0000FF', 0.9)).toBeFalsy();
+		expect(isLight('#FF00FF', 0.9)).toBeFalsy();
+		
+		expect(isLight('#FF0000', 0.1)).toBeTruthy();
+		expect(isLight('#FFFF00', 0.1)).toBeTruthy();
+		expect(isLight('#00FF00', 0.1)).toBeTruthy();
+		expect(isLight('#00FFFF', 0.1)).toBeTruthy();
+		expect(isLight('#0000FF', 0.1)).toBeTruthy();
+		expect(isLight('#FF00FF', 0.1)).toBeTruthy();
+	})
 
 	it('isHexColorValid', () => {
 		expect(isHexColorValid(undefined)).toBeFalsy();